+++ src/Config.cc
@@ -1,6 +1,6 @@
 #include "PLand/Config.h"
 #include "ll/api/Config.h"
-#include "mod/MyMod.h"
+#include "mod/ModEntry.h"
 #include <filesystem>
 
 
@@ -10,7 +10,7 @@ namespace fs = std::filesystem;
 #define CONFIG_FILE "Config.json"
 
 bool Config::tryLoad() {
-    auto dir = my_mod::MyMod::getInstance().getSelf().getConfigDir() / CONFIG_FILE;
+    auto dir = mod::ModEntry::getInstance().getSelf().getConfigDir() / CONFIG_FILE;
 
     if (!fs::exists(dir)) {
         trySave();
@@ -22,7 +22,7 @@ bool Config::tryLoad() {
 }
 
 bool Config::trySave() {
-    auto dir = my_mod::MyMod::getInstance().getSelf().getConfigDir() / CONFIG_FILE;
+    auto dir = mod::ModEntry::getInstance().getSelf().getConfigDir() / CONFIG_FILE;
 
     bool status = ll::config::saveConfig(cfg, dir);